+++ src/Symfony/Component/Validator/CHANGELOG.md
@@ -5,6 +5,15 @@ CHANGELOG
 ---
  * Add the `normalizer` option to the `Unique` constraint
  * Add `Validation::createIsValidCallable()` that returns true/false instead of throwing exceptions
+ * Add optional `conditionExpression` to the `Count` constraint to count only items matching the given symfony expression language
+
+   ```php
+   use Symfony\Component\Validator\Constraints as Assert;
+
+   /**
+    * @Assert\Count(exactly=3, conditionExpression="42 < item.getValue()")
+    */
+   ```
 
 5.2.0
 -----